+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### 功能构建
 
+- 新增 spring-telqos 框架集成指令。
+  - com.dwarfeng.tmpstg.api.integration.springtelqos.TmpstgCommand。
+
+- 增加依赖。
+  - 增加依赖 `spring-telqos` 以应用其新功能，版本为 `2.0.0.a`。
+
 - 为项目增加 xsd 配置机制。
   - 增加 `META-INF/dwarfeng-tmpstg.xsd` 文件。
   - 增加 `com.dwarfeng.tmpstg.node.configuration.TemporaryStorageNamespaceHandlerSupport` 及对应的定义解析器。